Restructure /admin around a ?tab= switcher
The page previously stacked all admin cards (Tokens, Backups, Merge
Candidates, Split Candidates) vertically. With the planned Overview
and Jobs Log sections, this layout would have pushed everything off
the bottom of the screen. Switch to a URL-driven tab nav so each
section gets full real estate and adding tabs is a one-place change.

Three tabs at launch: Tokens (default — extracted from the 132-line
inline block into RegistrationTokensCard, also takes the chance to
swap "Expires <date>" for "Used <date> by <user>" once a token is
consumed), Curation (Merge + Split cards stacked), Backups. Unknown
?tab= values fall back to the default so a stale bookmark never
lands the admin on a blank page.

Tabs eager-render and stay mounted (visibility toggles via
class:hidden, not {#if} unmount) — admin sessions usually touch
several tabs in a row, so the one-time parallel-fetch cost on first
paint buys instant subsequent switches. No card polls, so persistent
mount adds no ongoing traffic.

AdminTabNav uses theme-aware tokens (text-primary / border-primary
active, text-white/60 inactive against the dark body gradient) so
selected state picks up the user's chosen theme color.
