+++ phsar/frontend/CLAUDE.md
@@ -9,6 +9,7 @@ Pages: home (`/`), login (`/login`), register (`/register`), search (`/search`),
 - `GET /health` endpoint returning `{status, version}` for Coolify liveness — deliberately does not probe backend (liveness should only check what restarting this container can fix)
 - `/library/add` — query-input + recent-additions panel; submitting POSTs to `/jobs/scrape` and the navbar bell takes over from there; restricted users see the page but the form is disabled
 - **Tab title convention**: every route declares its own `<svelte:head><title>…</title></svelte:head>` in the format `<Page> — Phsar` (em dash, "Phsar" title case). Page name first so users with several PHSAR tabs open can distinguish them in the tab strip; the favicon already conveys "this is PHSAR". Detail pages (anime/media) bind the title reactively to the resolved name and fall back to a generic label while loading. `app.html` only sets the default `PHSAR` so the very first paint isn't blank — every route is expected to override
+- **Admin tab navigation**: `/admin` uses a `?tab=` query param to switch between sections (`tokens`, `curation`, `backups`). The tab list, default, and validation live at the top of `routes/admin/+page.svelte`; the tab UI itself is `lib/components/admin/AdminTabNav.svelte`. Adding a tab requires updating three places: `AdminTabKey` in `lib/components/admin/types.ts`, the `TABS` array + content cascade in the page, and the rendering branch
 - **Detail-page back button**: anime/media routes render `<BackLink searchToken fromParam />` near the top. It decides between "Back to search" (when `?q=<token>` is present, from search results) and "Back to library" (when `?from=library`, e.g. recent-additions panel). New origins extend the `DetailOrigin` type union in `lib/utils/navigation.ts` AND the `target` switch in `BackLink.svelte`. `buildDetailHref()` propagates both flags on internal anime↔media jumps so a deep dive (library → anime → media) stays linkable back to the origin
 
 ## lib/components/
