+++ phsar/frontend/src/lib/components/admin/types.ts
@@ -1,0 +1,5 @@
+/** Tab keys for the admin page. Extending this requires:
+ *  - adding the new entry to TABS in routes/admin/+page.svelte
+ *  - rendering the corresponding content branch in that file
+ */
+export type AdminTabKey = 'tokens' | 'curation' | 'backups';
